requirements.txt
# ffmpeg-normalize==1.28.3
apache-airflow==2.10.2
boto3==1.35.27
deepgram-sdk==3.4.0
psycopg2==2.9.9
deepl==1.19.1
anthropic==0.34.2
openai==1.37.1
nltk==3.8.1
mutagen==1.47.0
cuid2==2.0.1
paramiko==3.5.0
scp==0.15.0
vastai==0.2.5
jellyfish==1.1.0